+++ src/main/java/com/nhn/pinpoint/profiler/context/SpanId.java
@@ -4,20 +4,20 @@ import java.util.Random;
 
 public class SpanId {
 
-    public static final int NULL = -1;
+    public static final long NULL = -1;
 
     private static final Random seed = new Random();
 
-    public static int newSpanId() {
-        int id = seed.nextInt();
+    public static long newSpanId() {
+        long id = seed.nextLong();
         while (id == NULL) {
             id = seed.nextInt();
         }
         return id;
     }
 
-    public static int nextSpanID(int spanId, int parentSpanId) {
-        int newId = newSpanId();
+    public static long nextSpanID(long spanId, long parentSpanId) {
+        long newId = newSpanId();
         while (newId == spanId || newId == parentSpanId) {
             newId = newSpanId();
         }
